+++ Src/Light/Light.hpp
@@ -10,9 +10,9 @@ namespace Aurora
     class Light
     {
     public:
-        virtual RGB_T<float> Sample(const Vector4D_T<float> &p, const Vector4D_T<float> &n) = 0;
+        virtual RGB_T<double> Sample(const Vector4D_T<double> &p, const Vector4D_T<double> &n) = 0;
         virtual void Set(const nlohmann::json &json) = 0;
-        virtual void Transform(const Matrix4X4_T<float> &m) = 0;
+        virtual void Transform(const Matrix4X4_T<double> &m) = 0;
     private:
     };
     extern std::vector<std::shared_ptr<Light>> LightList;
